EnforcerRule's selector to respect exclusion
