+++ dependencies/src/main/java/com/google/cloud/tools/opensource/dependencies/RepositoryUtility.java
@@ -96,21 +96,6 @@ public final class RepositoryUtility {
   private static final ImmutableSet<String> ALLOWED_REPOSITORY_URL_SCHEMES =
       ImmutableSet.of("file", "http", "https");
 
-  public static final DependencySelector nonTestDependencySelector =
-      new DependencySelector() {
-
-        @Override
-        public boolean selectDependency(Dependency dependency) {
-          return !"test".equals(dependency.getScope());
-        }
-
-        @Override
-        public DependencySelector deriveChildSelector(
-            DependencyCollectionContext dependencyCollectionContext) {
-          return this;
-        }
-      };
-
   private RepositoryUtility() {}
 
   /**
